feat(i18n): add compact warning texts; remove Chat.tsx contents
Add "compactWarning" localization entries to en_US and ja_JP to warn when the message count exceeds the compact threshold and recommend using /compact or creating a new session. Remove the large Chat.tsx implementation (file contents cleared) as part of a refactor to relocate and simplify chat UI/logic.
