+++ todo_update
@@ -1,0 +1,1 @@
+これまでのtodoを更新して、今作成したコンポーネントは完了としてマークします。そして、メインのChatコンポーネントの作成に移ります。
